+++ src/pages/profile.js
@@ -1,5 +1,7 @@
 import { Header, Footer } from "../components";
 import { CONST } from "../constants";
+import { render } from "../router/router";
+import { state } from "../state";
 
 export const ProfilePage = () => `
 <div id="root">
@@ -65,3 +67,51 @@ export const ProfilePage = () => `
     </div>
     </div>
 `;
+
+export const onRenderProfile = () => {
+  if (!state.loggedInUser) {
+    return render(CONST.pathname.login);
+  }
+  const profileForm = document.getElementById(CONST.profileForm.formId);
+  if (!profileForm) return;
+
+  // 전역 객체에 저장된 user profile 데이터를 form에 초기화
+  if (state.loggedInUser) {
+    const fieldIdList = [
+      CONST.profileForm.field.username,
+      CONST.profileForm.field.bio,
+      CONST.profileForm.field.email,
+    ];
+    for (const fieldId of fieldIdList) {
+      const field = profileForm.querySelector(`#${fieldId}`);
+      field.value = state.loggedInUser[fieldId];
+    }
+  }
+
+  // 제출시 form의 data를 localstorage에 저장
+  profileForm.addEventListener("submit", (e) => {
+    e.preventDefault();
+    const formData = new FormData(profileForm);
+    const newData = Object.fromEntries(formData);
+    const oldData = state.loggedInUser;
+
+    const isProfileChanged =
+      newData.username !== oldData.username ||
+      newData.email !== oldData.email ||
+      newData.bio !== oldData.bio;
+
+    if (isProfileChanged) {
+      const newUserInfo = { ...state.loggedInUser, ...newData };
+      const userIndex = state.users.findIndex(
+        (user) => user.id === state.loggedInUser.id,
+      );
+      state.users.splice(userIndex, 1, newUserInfo);
+      localStorage.setItem(CONST.lsKey.users, JSON.stringify(state.users));
+
+      state.loggedInUser = newUserInfo;
+      localStorage.setItem(CONST.lsKey.user, JSON.stringify(newUserInfo));
+
+      alert("profile 변경 완료");
+    }
+  });
+};
